using first reducer in redux
