+++ src/App.js
@@ -1,11 +1,33 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import {
+  createUserDocumentFromAuth,
+  onAuthStateChangeListener,
+  // signOutUser,
+} from '../src/utils/firebase-utils/firebase.utils';
+
 import { Routes, Route } from 'react-router-dom';
 import Home from './routes/home/home.component';
 import Navigation from './routes/navigation/navigation.compoentn';
 import Authentication from './routes/authentication/authentication.component';
 import Shop from './routes/shop/shop.component';
 import Checkout from './routes/checkout/checkout.component';
+import { setCurrentUser } from './store/user/user.action';
 
 const App = () => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChangeListener((user) => {
+      if (user) {
+        createUserDocumentFromAuth(user);
+      }
+      dispatch(setCurrentUser(user));
+    });
+
+    return unsubscribe;
+  }, []);
+
   return (
     <Routes>
       <Route path="/" element={<Navigation />}>
@@ -13,8 +35,8 @@ const App = () => {
         <Route path="shop/*" element={<Shop />} />
         <Route path="auth" element={<Authentication />} />
         <Route path="checkout" element={<Checkout />} />
-    </Route>
-  </Routes>
+      </Route>
+    </Routes>
   );
 };
