+++ src/contex/user.contex.jsx
@@ -1,10 +1,4 @@
-import { createContext, useEffect, useReducer } from 'react';
-import {
-  createUserDocumentFromAuth,
-  onAuthStateChangeListener,
-  // signOutUser,
-} from '../utils/firebase-utils/firebase.utils';
-import { createAction } from '../utils/reducer/reducer.utils';
+import { createContext } from 'react';
 
 // as the actual value you want to access
 export const UserContext = createContext({
@@ -12,63 +6,10 @@ export const UserContext = createContext({
   setCurrentUser: () => null,
 });
 
-export const USER_ACTION_TYPES = {
-  SET_CURRENT_USER: 'SET_CURRENT_USER',
-};
-
-const userReducer = (state, action) => {
-  console.log('dispatch');
-  console.log(action);
-  const { type, payload } = action;
-
-  switch (type) {
-    case USER_ACTION_TYPES.SET_CURRENT_USER:
-      return {
-        ...state,
-        currentUser: payload,
-      };
-
-    default:
-      throw new Error(`Unhandled type ${type} in userReducer`);
-  }
-};
-
-const INITIAL_STATE = {
-  currentUser: null,
-};
-
 export const UserProvider = ({ children }) => {
-  // const [currentUser, setCurrentUser] = useState(null);
-
-  const [state, dispatch] = useReducer(userReducer, INITIAL_STATE);
   const { currentUser } = state;
 
-  const setCurrentUser = (user) => {
-    dispatch(createAction(USER_ACTION_TYPES.SET_CURRENT_USER, user));
-  };
   const value = { currentUser, setCurrentUser };
-
-  console.log(`currentUser : ${currentUser}`);
-
-  // signOutUser();
-  useEffect(() => {
-    const unsubscribe = onAuthStateChangeListener((user) => {
-      if (user) {
-        createUserDocumentFromAuth(user);
-      }
-      setCurrentUser(user);
-    });
-    return unsubscribe;
-  }, []);
 
   return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
 };
-
-/*
-  const userReducer = (state, action)=>{
-    return{
-      currentUser:{}
-    }
-  }
-
-*/
